reorganize channel code architecture & add dpa channel
