+++ examples/udp_echo/src/dpa_kernel.c
@@ -1,5 +1,4 @@
 #include "dpa_kernel.h"
-// #include "dpa_common.h"
 
 /*
 * This is the main function of the L2 reflector device, called on each packet from dpa_event_handler()
@@ -27,13 +26,13 @@ process_packet(struct flexio_dev_thread_ctx *dtctx)
   data_sz = flexio_dev_cqe_get_byte_cnt(dev_ctx.rqcq_ctx.cqe);
 
   /* Get RQ WQE pointed by CQE */
-  rwqe = &dev_ctx.rq_ctx.rq_ring[rq_wqe_idx & L2_RQ_IDX_MASK];
+  rwqe = &dev_ctx.rq_ctx.rq_ring[rq_wqe_idx & DPA_RQ_IDX_MASK];
 
   /* Extract data (whole packet) pointed by RQ WQE */
   rq_data = flexio_dev_rwqe_get_addr(rwqe);
 
   /* Take next entry from data ring */
-  sq_data = get_next_dte(&dev_ctx.dt_ctx, L2_DATA_IDX_MASK, L2_LOG_WQ_DATA_ENTRY_BSIZE);
+  sq_data = get_next_dte(&dev_ctx.dt_ctx, DPA_DATA_IDX_MASK, DPA_LOG_WQ_DATA_ENTRY_BSIZE);
 
   /* Copy received packet to sq_data as is */
   memcpy(sq_data, rq_data, data_sz);
@@ -47,22 +46,22 @@ process_packet(struct flexio_dev_thread_ctx *dtctx)
   }
 
   /* Take first segment for SQ WQE (3 segments will be used) */
-  swqe = get_next_sqe(&dev_ctx.sq_ctx, L2_SQ_IDX_MASK);
+  swqe = get_next_sqe(&dev_ctx.sq_ctx, DPA_SQ_IDX_MASK);
 
   /* Fill out 1-st segment (Control) */
   flexio_dev_swqe_seg_ctrl_set(swqe, dev_ctx.sq_ctx.sq_pi, dev_ctx.sq_ctx.sq_number,
             MLX5_CTRL_SEG_CE_CQE_ON_CQE_ERROR, FLEXIO_CTRL_SEG_SEND_EN);
 
   /* Fill out 2-nd segment (Ethernet) */
-  swqe = get_next_sqe(&dev_ctx.sq_ctx, L2_SQ_IDX_MASK);
+  swqe = get_next_sqe(&dev_ctx.sq_ctx, DPA_SQ_IDX_MASK);
   flexio_dev_swqe_seg_eth_set(swqe, mss, checksum, 0, NULL);
 
   /* Fill out 3-rd segment (Data) */
-  swqe = get_next_sqe(&dev_ctx.sq_ctx, L2_SQ_IDX_MASK);
+  swqe = get_next_sqe(&dev_ctx.sq_ctx, DPA_SQ_IDX_MASK);
   flexio_dev_swqe_seg_mem_ptr_data_set(swqe, data_sz, dev_ctx.lkey, (uint64_t)sq_data);
 
   /* Send WQE is 4 WQEBBs need to skip the 4-th segment */
-  swqe = get_next_sqe(&dev_ctx.sq_ctx, L2_SQ_IDX_MASK);
+  swqe = get_next_sqe(&dev_ctx.sq_ctx, DPA_SQ_IDX_MASK);
 
   /* Ring DB */
   __dpa_thread_fence(__DPA_MEMORY, __DPA_W, __DPA_W);
