+++ lib/meson.build
@@ -16,8 +16,7 @@ configure_file(input : root_dir + '/lib/log/debug.h.in', output : 'debug.h', con
 
 # compile targets
 subdir('log')
-subdir('wrapper/soc')
-subdir('channel')
+# subdir('wrapper/soc')
 
 # build nicc library
 message('source files: ' + ' '.join(lib_sources))
