+++ runtime/include/ctrlpath/mat.h
@@ -10,7 +10,7 @@
 #include "common.h"
 #include "log.h"
 #include "types.h"
-#include "utils/crc32.h"
+#include "libutils/crc32.h"
 
 
 namespace nicc
